[REF] resource: move HOURS_PER_DAY
Move the constant HOURS_PER_DAY back to where it was before, it is
business dependent and is not needed in tools.
Additionally, ROUNDING_FACTOR is never used.
