+++ addons/hr_holidays/models/hr_employee_base.py
@@ -4,8 +4,9 @@ from datetime import date, timezone, timedelta
 
 from odoo import api, fields, models, _
 from odoo.exceptions import ValidationError
-from odoo.tools.date_intervals import HOURS_PER_DAY
 from odoo.tools.float_utils import float_round
+
+from odoo.addons.resource.models.utils import HOURS_PER_DAY
 
 
 class HrEmployeeBase(models.AbstractModel):
