+++ addons/hr_holidays/wizard/hr_leave_allocation_generate_multi_wizard.py
@@ -1,7 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import _, api, fields, models
-from odoo.tools.date_intervals import HOURS_PER_DAY
+from odoo.addons.resource.models.utils import HOURS_PER_DAY
 
 
 class HrLeaveAllocationGenerateMultiWizard(models.TransientModel):
